# Fix bufio: buffer full error while unmarshaling large messages