--- a/decode.go
+++ b/decode.go
@@ -1,7 +1,6 @@
 package protocol
 
 import (
-	"bufio"
 	"bytes"
 	"encoding/binary"
 	"fmt"
@@ -231,47 +230,54 @@ type CommandDecoder interface {
 
 // JSONCommandDecoder ...
 type JSONCommandDecoder struct {
-	reader    *bytes.Reader
-	bufReader *bufio.Reader
+	//reader     *bytes.Reader
+	//bufReader  *bufio.Reader
+	decoder    *json.Decoder
+	data       []byte
+	isMultiple bool
 }
 
 // NewJSONCommandDecoder ...
 func NewJSONCommandDecoder(data []byte) *JSONCommandDecoder {
-	reader := bytes.NewReader(data)
+	//reader := bytes.NewReader(data)
+	isMultiple := bytes.Contains(data, []byte("\n"))
+	var decoder *json.Decoder
+	if isMultiple {
+		decoder = json.NewDecoder(bytes.NewReader(data))
+	}
 	return &JSONCommandDecoder{
-		reader:    reader,
-		bufReader: bufio.NewReader(reader),
+		//reader:     reader,
+		//bufReader:  bufio.NewReaderSize(reader, len(data)),
+		data:       data,
+		decoder:    decoder,
+		isMultiple: bytes.Contains(data, []byte("\n")),
 	}
 }
 
 // Reset ...
 func (d *JSONCommandDecoder) Reset(data []byte) error {
-	d.reader.Reset(data)
-	d.bufReader.Reset(d.reader)
+	isMultiple := bytes.Contains(data, []byte("\n"))
+	var decoder *json.Decoder
+	if isMultiple {
+		decoder = json.NewDecoder(bytes.NewReader(data))
+	}
+	d.data = data
+	d.isMultiple = isMultiple
+	d.decoder = decoder
 	return nil
 }
 
 // Decode ...
 func (d *JSONCommandDecoder) Decode() (*Command, error) {
 	var c Command
-	line, err := d.bufReader.ReadSlice(byte('\n'))
-	if err != nil {
-		if err == io.EOF {
-			if len(line) == 0 {
-				return nil, io.EOF
-			}
-			_, err = json.Parse(line, &c, json.ZeroCopy)
-			if err != nil {
-				return nil, err
-			}
-			return &c, io.EOF
+	if !d.isMultiple {
+		_, err := json.Parse(d.data, &c, json.ZeroCopy)
+		if err != nil {
+			return nil, err
 		}
-		return nil, err
+		return &c, io.EOF
 	}
-	if len(line) == 0 {
-		return nil, io.EOF
-	}
-	_, err = json.Parse(line, &c, json.ZeroCopy)
+	err := d.decoder.Decode(&c)
 	return &c, err
 }