Before:

```
BenchmarkReplyJSONUnmarshal-12            	 2456772	       487.5 ns/op	     232 B/op	       3 allocs/op
BenchmarkReplyJSONUnmarshalParallel-12    	10671240	       113.2 ns/op	     232 B/op	       3 allocs/op
```

After:

```
BenchmarkReplyJSONUnmarshal-12            	 2625344	       445.3 ns/op	     232 B/op	       3 allocs/op
BenchmarkReplyJSONUnmarshalParallel-12    	11780377	       102.6 ns/op	     232 B/op	       3 allocs/op
```

But for multiple messages we now fallback to decoder - thus performance will be slower. This should not be a very frequent case – but will try to address later. Currently fixing a bug and keeping performance the same (actually a bit faster) for most used scenario.